requirements.txt
charset-normalizer==3.4.4
colorama==0.4.6
iniconfig==2.3.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.1.1
reportlab==4.4.4
